+++ cpp/src/slice2cs/Main.cpp
@@ -6,6 +6,7 @@
 #include "../Slice/FileTracker.h"
 #include "../Slice/Preprocessor.h"
 #include "../Slice/Util.h"
+#include "CsUtil.h"
 #include "Gen.h"
 #include "Ice/CtrlCHandler.h"
