+++ cpp/src/slice2cs/msbuild/slice2cs.vcxproj
@@ -108,18 +108,22 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="..\Gen.cpp" />
     <ClCompile Include="..\CsMetadataValidator.cpp" />
-    <ClCompile Include="..\Main.cpp" />
     <ClCompile Include="..\CsUtil.cpp" />
+    <ClCompile Include="..\CsVisitor.cpp" />
+    <ClCompile Include="..\Gen.cpp" />
+    <ClCompile Include="..\IceVisitors.cpp" />
+    <ClCompile Include="..\Main.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\Slice2Cs.rc" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\Gen.h" />
     <ClInclude Include="..\CsMetadataValidator.h" />
     <ClInclude Include="..\CsUtil.h" />
+    <ClInclude Include="..\CsVisitor.h" />
+    <ClInclude Include="..\Gen.h" />
+    <ClInclude Include="..\IceVisitors.h" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\IceUtil\msbuild\iceutil\iceutil.vcxproj">
